fix(channels): every registered transport is published, and the load survives a bare server
Three corrections, two of them decisions rather than defects.

EVERY registered transport is published, core and unit alike. The security
review asked whether a unit's provider id — an operator's choice of what to
install — should reach every authenticated seat, and the answer is yes, decided
rather than inherited: a member whose timeline shows a message that arrived on a
unit's channel needs to know what to call it, and withholding the name leaves
that row rendering a raw id for exactly the transports the extension tier exists
to add. What is gated is the ACT, not the NAME — whether a member may send is an
RBAC question answered by the unit's own ext_<unit>_<object> grants and by the
send pre-flight. Disclosure and capability are separate axes, which is the same
separation this arc is about.

The directory load moves off newServer and onto the real assembly path. Route
level tests construct that one directly with a pool that was never dialled, so a
database read in a struct constructor panicked them — and a constructor is the
wrong place to reach a database regardless. Every role that serves /v1 comes
through New.

The channel vocabulary moves into its own file. It is a different question with a
different lifetime from the send path: the kind is fixed by the contract, the
transport set is decided at boot by what the binary composed, and ADR-0107/A158
is the decision that stopped one standing in for the other. The split also
returns both files under the 500-line ceiling.

Migration renumbered 0252 -> 0257: origin/main took 0252 through 0256 while this
branch was open. The two comments that named the number are now written without
one, which is what this repo's own trap list says to do.
